Third time's the charm, god whiling
The first iteration started at the top as a web-app and worked
downwards to a command line and core.

The second iteration started in the middle with a command-line and
worked down to a core.

Now we're starting at the bottom building a modular core library that
takes into account the design considerations of the previous attempts
and may be distributed standalone to add tagging and search to any
other CL program.