Specify the package when changing class in datum's initialize-instance
The previous behavior manifested as a bug when the data returned by a
get-datum call from the leibowitz-web package had the wrong class,
datum rather than datum-text.  Promoting the innermost class-not-found
error revealed that the hunchentoot handler method making the call was
searching for the class in cl-user rather than leibowitz-core.  What's
weird to me about this is that this bug didn't manifest sooner from
the tests.  I'm really not sure why find-class was searching cl-user
rather than leibowitz-web, which imports the symbols of
leibowitz-core, or why the test cases in leibowitz-core/tests (which
also imports leibowitz-core) are able to pass.  Am I unintentionally
running part of the web routing logic in cl-user?