added datum-html-preview method for link data
while I'm at it I might want to refactor the tree of datum types so
that the ones that correspond 1-1 to files on disk inherit from
datum-file.  having so many top-level datum methods assume they're
working with a single file on disk makes datum-link feel like a
second-class citizen and I worry it's caused some bugs to leak in to
the web and cli packages.